pyproject.toml
[project]
name = "cag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.90.0",
    "pymupdf>=1.26.1",
    "pymupdf4llm>=0.0.25",
    "python-magic>=0.4.27",
]

[build-system]
requires = ['setuptools']
build-backend = 'setuptools.build_meta'
